requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
supabase==2.0.0
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.4
email-validator==2.2.0
requests==2.31.0